+++ experimental/bindings/java/build.gradle
@@ -25,7 +25,7 @@ android {
 
         externalNativeBuild {
             cmake {
-                arguments '-DIREE_BUILD_JAVA_BINDINGS=ON',
+                arguments '-DIREE_BUILD_EXPERIMENTAL_JAVA_BINDINGS=ON',
                         '-DIREE_HAL_DRIVERS_TO_BUILD=VMLA',
 
                         // Disable all but the runtime components needed for the
@@ -35,21 +35,27 @@ android {
                         '-DIREE_BUILD_TESTS=OFF',
                         '-DIREE_BUILD_SAMPLES=OFF',
                         '-DIREE_BUILD_PYTHON_BINDINGS=OFF',
+
+                        // Run the commands at
+                        // https://google.github.io/iree/get-started/getting-started-android-cmake#configure-and-build
+                        // and update this flag to point to iree-build-host/install/.
+                        '-DIREE_HOST_BINARY_ROOT='
             }
         }
     }
 
     sourceSets {
         main {
-            java.srcDirs = ['com/google/iree']
-            jni.srcDirs = ['com/google/iree/native']
-            manifest.srcFile '../javatests/com/google/iree/TestManifest.xml'
+            java.srcDirs = ['java/com/google/iree']
+            jni.srcDirs = ['java/com/google/iree/native']
+            manifest.srcFile 'com/google/iree/tests/TestManifest.xml'
         }
     }
 
     externalNativeBuild {
         cmake {
-            path "../../CMakeLists.txt"
+            version "3.13.0+"
+            path "../../../CMakeLists.txt"
         }
     }
 }
