+++ experimental/bindings/java/com/google/iree/native/instance_jni.cc
@@ -14,7 +14,7 @@
 
 #include <jni.h>
 
-#include "bindings/java/com/google/iree/native/instance_wrapper.h"
+#include "experimental/bindings/java/com/google/iree/native/instance_wrapper.h"
 #include "iree/base/logging.h"
 
 #define JNI_FUNC extern "C" JNIEXPORT
